refactor: move tool definitions to separate file
