+++ src/index.ts
@@ -1,98 +1,14 @@
 import { Coinbase, Wallet } from "@coinbase/coinbase-sdk";
 import { Server } from "@modelcontextprotocol/sdk/server/index.js";
-import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
 import {
   CallToolRequestSchema,
   ListToolsRequestSchema,
-  type Tool,
 } from "@modelcontextprotocol/sdk/types.js";
 import { version } from "./version.js";
 import * as dotenv from "dotenv";
 import type { DeployContractParams, TransferFundsParams } from "./types.js";
-
-interface CreateWalletArgs {
-  seedPhrase: string;
-}
-
-// Tool definitions
-const getAddressTool: Tool = {
-  name: "get-address",
-  description: "Get the address for the wallet",
-  inputSchema: {
-    type: "object",
-  },
-};
-
-const getTestnetEthTool: Tool = {
-  name: "get-testnet-eth",
-  description:
-    "Get the testnet ETH balance for the wallet. Can only be called on Base Sepolia",
-  inputSchema: {
-    type: "object",
-  },
-};
-
-const listBalancesTool: Tool = {
-  name: "list-balances",
-  description: "List all balances for a wallet",
-  inputSchema: {
-    type: "object",
-  },
-};
-
-const transferFundsTool: Tool = {
-  name: "transfer-funds",
-  description: "Transfer funds from one wallet to another",
-  inputSchema: {
-    type: "object",
-    properties: {
-      destination: {
-        type: "string",
-        description: "The address to which to transfer funds",
-      },
-      assetId: {
-        type: "string",
-        enum: Object.values(Coinbase.assets),
-        description: "The asset ID to transfer",
-      },
-      amount: {
-        type: "number",
-        description: "The amount of funds to transfer",
-      },
-    },
-  },
-};
-
-const deployContractTool: Tool = {
-  name: "deploy-contract",
-  description: "Deploy a contract",
-  inputSchema: {
-    type: "object",
-    properties: {
-      constructorArgs: {
-        type: "array",
-        description: "The arguments for the contract constructor",
-        items: {
-          type: "string",
-        },
-      },
-      contractName: {
-        type: "string",
-        description: "The name of the contract to deploy",
-      },
-      solidityInputJson: {
-        type: "string",
-        description:
-          "The JSON input for the Solidity compiler containing contract source and settings",
-      },
-      solidityVersion: {
-        type: "string",
-        description: "The version of the solidity compiler",
-      },
-    },
-  },
-};
+import { tools } from "./tools/index.js";
 
 async function main() {
   dotenv.config();
@@ -130,13 +46,7 @@ async function main() {
   server.setRequestHandler(ListToolsRequestSchema, async () => {
     console.error("Received ListToolsRequest");
     return {
-      tools: [
-        getAddressTool,
-        listBalancesTool,
-        getTestnetEthTool,
-        transferFundsTool,
-        deployContractTool,
-      ],
+      tools,
     };
   });
